fix bag constants, small refactoring

+++ constants/constants.h
@@ -21,7 +21,7 @@ struct lic_host_status {
 
 struct constants {
 #define _CONSTANTS
-#define DAT(a, b, c) const static int a = b;
+#define DAT(a, b, c) const static long a = b;
 #include "constants.dat"
 #undef _CONSTANTS
 #undef DAT
